asset_pipeline/requirements_3dsmax.txt
attrs==22.2.0
bddl==3.0.0b1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.6
decorator==5.1.1
distlib==0.3.4
exceptiongroup==1.1.0
fastjsonschema==2.16.2
filelock==3.6.0
freetype-py==2.3.0
fs==2.4.16
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==5.3.2
idna==3.3
imageio==2.22.4
importlib-metadata==4.11.3
importlib-resources==5.10.2
iniconfig==2.0.0
jsonschema==4.17.3
jupyter-core==4.12.0
jupytext==1.14.4
lxml==4.9.1
markdown-it-py==2.1.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
nbformat==5.7.3
networkx==2.2
nltk==3.8.1
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==23.0
pandas==1.3.5
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycollada==0.6
pyglet==2.0.0
PyOpenGL==3.1.0
pyrender==0.1.45
pyrsistent==0.19.3
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.1
pywin32==305
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
six==1.16.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
traitlets==5.8.1
trimesh==3.11.2
typing-extensions==4.2.0
urllib3==1.26.9
virtualenv==20.14.1
zipp==3.8.0
rtree==1.0.1